+++ dev-kit/exportmlx_gendoco.m
@@ -16,9 +16,7 @@ docFiles = [
 for fileStem = docFiles(:)'
     mlxFile = fullfile(docDir, fileStem + ".mlx");
     fprintf('Exporting ExportMlx doc file: %s\n', fileStem + ".mlx");
-    janklab.exportmlx.mlx2latex(mlxFile);
-    noExtnFile = fullfile(docDir, fileStem);
-    janklab.exportmlx.lslatex2markdown(noExtnFile);
+    janklab.exportmlx.livescript2markdown(mlxFile);
 end
 
 % And all the examples
@@ -27,12 +25,9 @@ examplesDir = fullfile(rootDir, 'examples');
 d = dir(examplesDir + "/*.mlx");
 exampleMlxs = string({d.name});
 for mlxFileBase = exampleMlxs
-    [~, fileStem, ~] = fileparts(mlxFileBase);
     mlxFile = fullfile(examplesDir, mlxFileBase);
     fprintf('Exporting ExportMlx example file: %s\n', mlxFileBase);
-    janklab.exportmlx.mlx2latex(mlxFile);
-    noExtnFile = fullfile(examplesDir, fileStem);
-    janklab.exportmlx.lslatex2markdown(noExtnFile);    
+    janklab.exportmlx.livescript2markdown(mlxFile);
 end
 
 fprintf('Exported ExportMlx doco.\n')